Binary glm object at several threshold values

dr_asympar computes a normal approximation of the likelihood at a vector of threshold values

dr_asympar(y, x, thresh, ...)

Arguments

y
outcome variable

x
matrix of covariates

thresh
vector of threshold values on the support of outcome y

...
additional arguments to pass to lapl_aprx2

Value

a list of glm objects corresponding to thresh

Examples

y = faithful$waiting
x = scale(cbind(faithful$eruptions,faithful$eruptions^2))
qtaus = quantile(y,c(0.05,0.25,0.5,0.75,0.95))
drabj<- dr_asympar(y=y,x=x,thresh = qtaus)
lapply(drabj,coef); lapply(drabj,vcov)

#> $`5%`
#> (Intercept)          X1          X2 
#>  -12.529504    7.552197  -16.412033 
#> 
#> $`25%`
#> (Intercept)          X1          X2 
#>   -6.222253    4.537972  -10.468049 
#> 
#> $`50%`
#> (Intercept)          X1          X2 
#>    2.048402  -13.100002    8.855666 
#> 
#> $`75%`
#> (Intercept)          X1          X2 
#>    2.522160   -6.544941    4.100806 
#> 
#> $`95%`
#> (Intercept)          X1          X2 
#>    8.356307  -21.168439   13.933436 
#> 

#> $`5%`
#>             (Intercept)        X1        X2
#> (Intercept)    295.8311 -401.3021  648.1783
#> X1            -401.3021  568.4650 -904.5508
#> X2             648.1783 -904.5508 1446.7640
#> 
#> $`25%`
#>             (Intercept)        X1        X2
#> (Intercept)    8.957952 -16.91239  24.55477
#> X1           -16.912395  40.88275 -55.52501
#> X2            24.554773 -55.52501  76.72885
#> 
#> $`50%`
#>             (Intercept)         X1         X2
#> (Intercept)   0.9382644  -5.362741   4.168658
#> X1           -5.3627412  33.586619 -26.632828
#> X2            4.1686583 -26.632828  21.260327
#> 
#> $`75%`
#>             (Intercept)        X1         X2
#> (Intercept)   0.3778717  -1.93365   1.464948
#> X1           -1.9336498  13.05132 -10.482798
#> X2            1.4649478 -10.48280   8.539182
#> 
#> $`95%`
#>             (Intercept)        X1        X2
#> (Intercept)    36.33379 -165.7202  120.2247
#> X1           -165.72018  773.3070 -564.5566
#> X2            120.22469 -564.5566  412.9514
#> 

# mean and covariance at respective threshold values