Joint asymptotic mutivariate density of parameters

jdpar.asymp takes input object from dr_asympar() for asymptotic bayesian distribution. It returns objects for joint mutivariate density of parameters across several thresholds. Check for positive definiteness of the covariance matrix, else exclude thresholds yielding negative eigen values.

jdpar.asymp(drabj, data, jdF = FALSE, vcovfn = "vcovHC", ...)

Arguments

drabj
object from dr_asympar()

data
dataframe, first column is the outcome

jdF
logical to return joint density across thresholds in drabj

vcovfn
a string denoting the function to extract the variance-covariance. Defaults at "vcov". Other variance-covariance estimators in the sandwich package are usable.

...
additional input to pass to vcovfn

Value

mean vector Theta and variance-covariance matrix vcovpar of parameters across thresholds and if jdF=TRUE, a mean vector mnF and a variance-covariance matrix vcovF of F(yo)